added user service and exceptions

+++ FreelancePlatform/src/main/java/freelanceplatform/data/FeedBackRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FeedBackRepository extends CrudRepository<Feedback, Long> {
+public interface FeedBackRepository extends CrudRepository<Feedback, Integer> {
 }
