+++ FreelancePlatform/src/main/java/freelanceplatform/data/ProposalRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProposalRepository extends CrudRepository<Proposal, Long> {
+public interface ProposalRepository extends CrudRepository<Proposal, Integer> {
 }
